+++ src/main/java/dev/pro/shelter/model/Users.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private Long chatId;
 
     @Embedded
@@ -22,7 +22,7 @@ public class Users {
 //    @JsonIgnore
 //    private List<Report> reports;
 
-    public Users(Integer id, Long chatId, Contact contact) {
+    public Users(Long id, Long chatId, Contact contact) {
         this.id = id;
         this.chatId = chatId;
         this.contact = contact;
@@ -31,11 +31,11 @@ public class Users {
     public Users() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
