+++ src/main/java/dev/pro/shelter/service/DogService.java
@@ -5,8 +5,8 @@ import dev.pro.shelter.model.Dog;
 
 public interface DogService extends PetService{
     Dog createDog(Dog dog);
-    Dog readDog(int id);
+    Dog readDog(long id);
     Dog updateDog(Dog dog);
-    Dog deleteDog(int id);
+    Dog deleteDog(long id);
 
 }
